refactor: entity 패키지 추가
entity 패키지에 관련 클래스들을 이동
