+++ src/main/kotlin/study/step3/entity/Scoreboard.kt
@@ -1,4 +1,4 @@
-package study.step3
+package study.step3.entity
 
 class Scoreboard {
     private val scoresList = mutableListOf<List<Score>>()
